Refactor processing - split up into multiple components
