+++ clients/gisquick-web/src/components/ogc-processes/OgcProcesses.vue
@@ -39,50 +39,16 @@
             <v-icon name="layers" size="14" class="mr-1"/>
             <span>Spatial layer added to map</span>
           </div>
-          <div
-            v-if="!result.artifacts || !result.artifacts.length"
+          <span
+            v-if="!result.ows_url && (!result.artifacts || !result.artifacts.length)"
             class="no-artifacts"
           >
-            <span v-if="!result.ows_url">Process completed with no file outputs.</span>
-          </div>
-          <div v-else class="artifacts-list f-col">
-            <div
-              v-for="artifact in result.artifacts"
-              :key="artifact.output_id"
-              class="artifact-item f-col"
-            >
-              <div class="artifact-row f-row-ac">
-                <v-icon name="file-outline" size="14" class="mr-1 flex-shrink-0"/>
-                <span class="artifact-name f-grow">{{ artifact.path }}</span>
-                <span class="artifact-size">{{ formatSize(artifact.size_bytes) }}</span>
-                <a :href="artifact.download_url" :download="artifact.path">
-                  <v-btn class="icon small">
-                    <v-icon name="download" size="14"/>
-                  </v-btn>
-                </a>
-                <v-btn
-                  v-if="artifact.media_kind === 'json'"
-                  class="icon small"
-                  @click="toggleJsonArtifact(artifact)"
-                >
-                  <v-icon :name="expandedArtifacts[artifact.output_id] ? 'arrow-down' : 'arrow-right'" size="14"/>
-                </v-btn>
-              </div>
-              <div
-                v-if="artifact.media_kind === 'json' && expandedArtifacts[artifact.output_id]"
-                class="artifact-json"
-              >
-                <v-spinner v-if="loadingArtifacts[artifact.output_id]" size="16" width="2"/>
-                <json-viewer
-                  v-else-if="jsonArtifactData[artifact.output_id]"
-                  :data="jsonArtifactData[artifact.output_id]"
-                />
-                <span v-else-if="artifactErrors[artifact.output_id]" class="error-text">
-                  {{ artifactErrors[artifact.output_id] }}
-                </span>
-              </div>
-            </div>
-          </div>
+            Process completed with no file outputs.
+          </span>
+          <result-artifacts-list
+            v-if="result.artifacts && result.artifacts.length"
+            :artifacts="result.artifacts"
+          />
         </div>
       </div>
       <div v-if="executeError" class="error-panel mt-2 p-2">
@@ -96,29 +62,12 @@
 import axios from 'axios'
 import OgcProcessSelect from './OgcProcessSelect.vue'
 import OgcProcessForm from './OgcProcessForm.vue'
-import JsonViewer from './JsonViewer.vue'
-
-const OGC_REL_RESULTS = 'http://www.opengis.net/def/rel/ogc/1.0/results'
-
-function extractResultsUrl (links) {
-  if (!Array.isArray(links)) return null
-  const candidates = links.filter(l => l.rel === OGC_REL_RESULTS)
-  return candidates.find(l => l.type === 'application/json')?.href
-    || candidates.find(l => l.type === 'application/geo+json')?.href
-    || candidates[0]?.href
-    || null
-}
-
-const POLL_INTERVAL_MS = 2000
-const POLL_MAX_RETRIES = 300  // 300 × 2s = 10 minutes
-const FAILURE_STATUSES = new Set(['failed', 'error', 'dismissed', 'canceled', 'cancelled'])
-
-function sleep (ms) {
-  return new Promise(resolve => setTimeout(resolve, ms))
-}
+import ResultArtifactsList from './ResultArtifactsList.vue'
+import { extractResultsUrl, sleep } from './utils'
+import { POLL_INTERVAL_MS, POLL_MAX_RETRIES, FAILURE_STATUSES } from './schema'
 
 export default {
-  components: { OgcProcessSelect, OgcProcessForm, JsonViewer },
+  components: { OgcProcessSelect, OgcProcessForm, ResultArtifactsList },
   props: {
     baseUrl: {
       type: String,
@@ -137,14 +86,9 @@ export default {
       result: null,
       executeError: null,
       polling: false,
-      pollingStatus: '',
-      expandedArtifacts: {},
-      loadingArtifacts: {},
-      jsonArtifactData: {},
-      artifactErrors: {}
+      pollingStatus: ''
     }
   },
-  computed: {},
   watch: {
     selectedProcessId () {
       this.clearResult()
@@ -206,6 +150,7 @@ export default {
         this.pollingStatus = ''
       }
     },
+
     async execute () {
       this.$refs.processForm?.finalizeActivePicking()
       this.executing = true
@@ -236,33 +181,10 @@ export default {
         this.executing = false
       }
     },
-    async toggleJsonArtifact (artifact) {
-      const id = artifact.output_id
-      this.$set(this.expandedArtifacts, id, !this.expandedArtifacts[id])
-      if (!this.expandedArtifacts[id] || this.jsonArtifactData[id]) return
-      this.$set(this.loadingArtifacts, id, true)
-      try {
-        const { data } = await axios.get(artifact.download_url)
-        this.$set(this.jsonArtifactData, id, data)
-      } catch (err) {
-        this.$set(this.artifactErrors, id, err.message)
-      } finally {
-        this.$set(this.loadingArtifacts, id, false)
-      }
-    },
-    formatSize (bytes) {
-      if (!bytes) return ''
-      if (bytes < 1024) return `${bytes} B`
-      if (bytes < 1048576) return `${(bytes / 1024).toFixed(1)} KB`
-      return `${(bytes / 1048576).toFixed(1)} MB`
-    },
+
     clearResult () {
       this.result = null
       this.executeError = null
-      this.expandedArtifacts = {}
-      this.loadingArtifacts = {}
-      this.jsonArtifactData = {}
-      this.artifactErrors = {}
     }
   }
 }
@@ -299,39 +221,6 @@ export default {
     .no-artifacts {
       font-size: 0.85em;
       opacity: 0.6;
-    }
-    .artifacts-list {
-      gap: 4px;
-    }
-    .artifact-item {
-      border: 1px solid rgba(0,0,0,0.08);
-      border-radius: 4px;
-      overflow: hidden;
-    }
-    .artifact-row {
-      padding: 4px 6px;
-      gap: 4px;
-      background: rgba(0,0,0,0.02);
-    }
-    .artifact-name {
-      font-size: 0.85em;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    }
-    .artifact-size {
-      font-size: 0.75em;
-      opacity: 0.55;
-      white-space: nowrap;
-    }
-    .artifact-json {
-      padding: 6px 8px;
-      overflow: auto;
-      max-height: 300px;
-    }
-    .error-text {
-      color: var(--color-red, #d32f2f);
-      font-size: 0.85em;
     }
   }
   .error-panel {
